Roles usuarios

+++ app/Http/Controllers/Admin/PostsController.php
@@ -8,35 +8,42 @@ use App\Post;
 use App\Tag;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use illuminate\Support\Str;
-
+use Illuminate\Support\Str;
 
 class PostsController extends Controller
 {
     public function index()
     {
-        $posts = Post::all();
+        $posts = Post::allowed()->get();
         return view('admin.posts.index', compact('posts'));
     }
 
-    public function store(Request $request){
+    public function store(Request $request)
+    {
+        $this->authorize('create', new Post);
+
         $this->validate($request, ['title' => 'required | min:3']);
-        $post = Post::create([
-            'title' =>  $request->title,
-            'user_id' => auth()->user()->id
-        ]);
+
+        $post = Post::create($request->all());
+
         return redirect()->route('admin.posts.edit', $post);
     }
 
     public function edit(Post $post)
     {
-        $categories = Category::all();
-        $tags = Tag::all();
-        return view('admin.posts.edit', compact('post','categories','tags'));
+        $this->authorize('view', $post);
+
+        return view('admin.posts.edit', [
+            'post' => $post,
+            'categories' => Category::all(),
+            'tags' => Tag::all()
+        ]);
     }
 
     public function update(StorePostRequest $request, Post $post)
     {
+        $this->authorize('update',$post);
+
         $post->update($request->all());
 
         $post->syncTags($request->tags);
@@ -46,9 +53,12 @@ class PostsController extends Controller
             ->with('flash', 'El post ha sido actualizado correctamente');
     }
 
-    public function destroy (Post $post)
+    public function destroy(Post $post)
     {
+        $this->authorize('delete', $post);
+
         $post->delete();
+
         return redirect()->route('admin.posts.index')->with('flash', 'El post ha sido eliminado');
     }
 }
